Extract SimpleSystemAudioDevice extension
